play nicer with smartparens
xenops-mode remaps `)` to `xenops-math-insert-closing-paren`, which as part of
its execution uses `insert` to add a closing paren. The problem is that this
confuses smartparens, which usually tries to skip the closing pair.

For example, suppose I am at a blank line and have both smartparens-mode and
xenops-mode enabled. I strike `(` and now my line looks like, where point is
indicated by `|`

(|)

If I now strike `)`, I expect the line to look like this:

()|

However, it will instead look like this:

()|)

This is because smartparens only "swoops in" after a `self-insert-command` and
not after `insert`. To fix this, I've changed the way that xenops inserts the
closing `)` so that `smartparens` will still fire