resolve deferred properties
when xenops-mode parses a src block, it uses the `xenops-src-do-in-org-mode`
macro, which essentially tosses the contents into a temporary buffer and parses
it there. The problem is that org now allows many properties to be *deferred*,
in which case they are retrieved later from the relevant buffer. However, by the
time that happens, the temporary buffer has been discarded.

A workaround is to use `org-element-resolve-deferred` to do what it says on the
tin: resolve any deferred properties while the temporary buffer is still available